+++ api/dkml_install_api_intf.ml
@@ -303,7 +303,7 @@ module type Intf = sig
     let setup_log style_renderer level =
       Fmt_tty.setup_std_outputs ?style_renderer ();
       Logs.set_level level;
-      Logs.set_reporter (Logs_fmt.reporter ());
+      Logs.set_reporter (Logs.format_reporter ());
       Dkml_install_api.Log_config.create ?log_config_style_renderer:style_renderer
         ?log_config_level:level ()