+++ runner/src/cmdliner_runner.ml
@@ -33,7 +33,7 @@ let help_secs =
 let setup_log style_renderer level =
   Fmt_tty.setup_std_outputs ?style_renderer ();
   Logs.set_level level;
-  Logs.set_reporter (Logs_fmt.reporter ());
+  Logs.set_reporter (Logs.format_reporter ());
   Dkml_install_api.Log_config.create ?log_config_style_renderer:style_renderer
     ?log_config_level:level ()